Drop exec, pause, resume, and signal
All of these require state information to be shared between funC
invocations (to map from a container ID to the
cgroups/namespaces/etc.), and after today's meeting we may be backing
away from that [1,2].  Even if we keep a requirement for sharing state
between funC invocations, we don't want to specify these IPC-requiring
commands until we have more clarity on that requirement in the spec.

On systems like Solaris, the kernel maintains a registry of container
IDs directly, so they don't need an external registry [3].  But
without a consensus around the minimal amount of inter-process state
sharing, we don't want to require container ID → state lookups in the
command-line spec.  Once we have more clarity on a minimal required
mechanism (e.g. Julz's --state-file [4,5,6]), we can add them back in
with an API that all runtimes can easily support (although runtimes
are of course free to provide more convenient APIs as additional
extensions).

Pause, resume, and signal are still in the current lifecycle pull
request [7], but I've requested they be removed until we have more
clarity around the basic lifecycle [8].

[1]: http://ircbot.wl.linuxfoundation.org/meetings/opencontainers/2015/opencontainers.2015-12-02-18.01.html
[2]: http://ircbot.wl.linuxfoundation.org/meetings/opencontainers/2015/opencontainers.2015-12-02-18.01.log.html#l-79
[3]: https://github.com/wking/oci-command-line-api/pull/3#issuecomment-162079383
[4]: https://github.com/wking/oci-command-line-api/pull/3#issuecomment-162009033
[5]: https://github.com/wking/oci-command-line-api/pull/3#issuecomment-162041009
[6]: https://github.com/wking/oci-command-line-api/pull/3#issuecomment-162082556
[7]: https://github.com/mrunalp/specs/commit/bd549a2c9569fd5d775323f7f39c835baf825d0a#diff-b84a8d65d8ed53f4794cd2db7e8ea731R48
[8]: https://github.com/opencontainers/specs/pull/231/files#r45532412

Signed-off-by: W. Trevor King <wking@tremily.us>
Reviewed-by: Mike Brown <brownwm@us.ibm.com>
Reviewed-by: Jesse Butler <jeeves.butler@gmail.com>
Reviewed-by: Julian Friedman <julz.friedman@uk.ibm.com>