Makefile: Add rootfs/bin/busybox rule
Based on [1] and relicenced under the Apache License, Version 2.0.

The 'sudo' for mkdir generates a directory belonging to root.

We check the digest signature and SHA512 hash of the stage3 to ensure
the downloaded file is from Gentoo's release engineering team.  For
the SHA512 check, the first grep removes the Whirlpool hashes to avoid
confusing sha512sum.  The second grep removes the '--' separator added
by the first grep.

The touch avoids attempted rebuilds if the stage3's busybox has an
mtime older than the local dependencies like
downloads/stage3-amd64-current.tar.bz2.

[1]: https://github.com/wking/oci-gentoo-minimal/blob/5f6850f05a4c183434cdb77774b10c9719649d49/Makefile#L22-L25

Signed-off-by: W. Trevor King <wking@tremily.us>