start: Change --config and --runtime to --bundle
Personally I prefer a single config file [1].  I want folks to be able
to pipe their config into the 'funC start' command (e.g. via a
/dev/fd/3 pseudo-filesystem path) [2], and I have a working example
that supports this without difficulty [3].  But since [4] landed on
2015-11-16, runC has replaced their --config-file and --runtime-file
flags with --bundle, and the current goal of this repository is
"keeping as much similarity with the existing runC command-line as
possible", not "makes sense to Trevor" ;).

[1]: https://groups.google.com/a/opencontainers.org/forum/#!topic/dev/0QbyJDM9fWY
     Subject: Single, unified config file (i.e. rolling back specs#88)
     Date: Wed, 4 Nov 2015 09:53:20 -0800
     Message-ID: <20151104175320.GC24652@odin.tremily.us>
[2]: https://github.com/opencontainers/runc/issues/202
[3]: https://github.com/wking/ccon/tree/8ab5b535b5eca1a62e12b5e865735e24f1e1666d#configuration
[4]: https://github.com/opencontainers/runc/pull/373

Signed-off-by: W. Trevor King <wking@tremily.us>
Reviewed-by: Mike Brown <brownwm@us.ibm.com>