CONTRIBUTING: Document Signed-off-by and the DCO
This project seems to have been operating under the assumption that
the usual Signed-off-by <-> DCO mapping applied.  But it's nice to
make that explicit.  This commit copies the current DCO section from
runtime-spec, but puts it in CONTRIBUTING.md for a more native GitHub
experience [1].

Looking through the history of the repo, the only non-merge commits
without sign-offs are by Mrunal:

* 5286bb06 (Update runtime validation section, 2016-01-19)
* b71f0452 (Update README.md, 2015-11-09)
* 2c0e9e20 (Update help for generate, 2015-11-05)
* 850abe9f (Update README.md, 2015-09-15)
* bb903899 (Initial commit, 2015-09-15)

who clearly doesn't have a problem signing for the rest of his work
and is unlikely to have stolen those contributions from somewhere else
;).

[1]: https://github.com/blog/1184-contributing-guidelines

Signed-off-by: W. Trevor King <wking@tremily.us>