translate: Stub in a translation framework
Add tooling to translate higher-level configs into the basic OCI
config.  On IRC, Julz floated a linux.namespaces[].fromContainer as a
higher-level version of linux.namespaces[].path for emulating exec
[1].  That makes sense to me, and Mrunal is open to something like
[2]:

  $ ocitools generate --template=high-level-config.json --translate=fromContainer --runtime=runc

The interface{} -> rspec.Spec conversion happens through a JSON
serialization trick suggested by 梁辰晔 (Liang Chenye) [3] after the
translations, because before the translations the template may contain
JSON that's not represented in the OCI Spec structure
(e.g. fromContainer in namespace entries).

The /proc lookup could still be improved:

* Lookup the mount path (currently hard-coded to /proc) in
  /proc/self/mounts.  This would extend support to users who had
  mounted proc at a different location, but it's hard to find
  self/mounts unless you already know where proc is mounted ;).
* Compare /proc with the state JSON's implied PID namespace [4].  This
  would extend support to users in a different PID namespace than the
  one in which the state JSON's 'pid' entry applied.

But those are likely to be corner cases (and until something like [4]
lands, the latter is not possible).  So I'm punting them into the
future, where they can be fixed as we discover methods to fix them.

[1]: http://ircbot.wl.linuxfoundation.org/eavesdrop/%23opencontainers/%23opencontainers.2016-04-27.log.html#t2016-04-27T20:32:09
[2]: http://ircbot.wl.linuxfoundation.org/eavesdrop/%23opencontainers/%23opencontainers.2016-04-28.log.html#t2016-04-28T16:12:48
[3]: https://github.com/opencontainers/ocitools/pull/54#issuecomment-216499298
[4]: https://groups.google.com/a/opencontainers.org/forum/#!topic/dev/ujtABQoCmgk
     Subject: Linux: Adding the PID namespace inode to state JSON
     Date: Wed, 30 Dec 2015 21:34:57 -0800
     Message-ID: <20151231053457.GG3680@odin.tremily.us>

Signed-off-by: W. Trevor King <wking@tremily.us>