generate: Don't add namespaces by default
And actively remove them if the PATH is the keyword "host".

This helps with --template, because it's harder to get clean template
round-trips if you have to opt-out of the namespace additions.  The
default template includes the non-user namespaces, so this commit
doesn't effect the bare 'ocitools generate' output.

Also document the PATH argument in ocitools-generate(1).  I couldn't
find a way to support '--network', etc. in codegangsta/cli, so the:

  --network=[PATH]

and such are intentional.  I would have preferred --network[=PATH],
but using --network= instead of --network for "create a new network
namespace" isn't that bad.

Signed-off-by: W. Trevor King <wking@tremily.us>