CONTRIBUTING: Make leader-issues optional
For small projects like ocitools a strict requirement would just be
busywork.  And the risk here (a submitted PR that would have been
rejected or redirected in a leader issue) is a risk assumed by the
submitter (who sunk time in the wrong direction) and not much cost to
the maintainers.

Signed-off-by: W. Trevor King <wking@tremily.us>