CONTRIBUTING: Allow collaborative pull requests
For runtime-spec, there are often PRs that pickup and reroll another
user's commits (e.g. opencontainers/runtime-spec#337). As long as the
Signed-off-by entries are there (for the DCO) and the new PR
references the earlier work (to avoid maintainer confusion), I see no
problem with this sort of collaboration.

I thought about replacing the old wording with words like the above
paragraph, but it seemed overly prescriptive.

Signed-off-by: W. Trevor King <wking@tremily.us>