proposals: release approval process: use consistent language for rejects
Fix up the language around REJECTs so it is easier to understand. The
basic premise is that a release may continue with REJECTs if 2/3 of the
maintainers vote to make the release. But, the maintainers SHOULD
discuss and allow time for any REJECTs to become LGTMs. Spread over two
discussions:

[1](https://github.com/opencontainers/tob/pull/15/files/bdfa70d70f093146bc730be2576586ec8ed57cca#r66519789)
and
[2](https://github.com/opencontainers/tob/pull/15/files/bdfa70d70f093146bc730be2576586ec8ed57cca#r66668148)