.mailmap: Add entries for inconsistent users
Where the same user had multiple entries, I mostly went with whichever
entry had the most non-merge commits.  Before this commit:

  $ git shortlog -se --no-merges
      ...
       5  Mrunal Patel <mrunal@me.com>
      81  Mrunal Patel <mrunalp@gmail.com>
      ...
       1  Wang Qilin <qilin.wang@huawei.com>
       5  Wang Qilin <wangqilin2@huawei.com>
       8  liang chenye <liangchenye@huawei.com>
       8  liangchenye <liangchenye@huawei.com>
      ...
       1  梁辰晔 (Liang Chenye) <liangchenye@huawei.com>

With Liang, I went with the version that also included the logograms,
because logograms are cool and it matches his GitHub name [1] ;).

The order is alphabetical according to Emacs' sort-lines ;).

Details on the format in git-shortlog(1).

[1]: https://github.com/liangchenye

Signed-off-by: W. Trevor King <wking@tremily.us>