runtime: Add a 'state' command
Partially catch up with opencontainers/runtime-spec@7117ede7 (Expand
on the definition of our ops, 2015-10-13,
opencontainers/runtime-spec#225, v0.4.0).  The state example is
adapted from the current release [1], but we defer the actual
definition of that JSON to runtime-spec.

The encoding for the output JSON (and all standard-stream activity) is
covered by the "Character encodings" section.  In cases where the
runtime ignores the SHOULD (still technically compliant), RFC 7159
makes encoding detection reasonably straightforward [2].  The obsolete
RFC 4627 has some hints as well [3] (although these were dropped in
RFC 7518 [4], probably as a result of removing the constraint that
"JSON text" be an object or array [4]).  The hints should still apply
to the state output, because we know it will be an object.  If that
ends up being too dicey and we want to certify runtimes that do not
respect their operating-system conventions, we can add an --encoding
option later.

[1]: https://github.com/opencontainers/runtime-spec/blob/v1.0.0-rc1/runtime.md#state
[2]: https://tools.ietf.org/html/rfc7159#section-8.1
[3]: https://tools.ietf.org/html/rfc4627#section-3
[4]: https://tools.ietf.org/html/rfc7158#appendix-A
     Which is currently identical to 7519.

Signed-off-by: W. Trevor King <wking@tremily.us>