command-line-interface: Rename from runtime.md to avoid collisions
runtime-spec already has a runtime.md talking about operations at an
interface-agnostic level.  Rename this file so that it's clearly
limited to the CLI.  runtime-tools doesn't have a runtime.md, but the
more-specific command-line-interface.md will make rebasing a merge
into runtime-tools easier if runtime-tools happens to grow a
runtime.md in the meantime.

Signed-off-by: W. Trevor King <wking@tremily.us>