*: clarify how security issues are handled
The security@opencontainers.org mailing list is for *maintainers only*,
and is to be used for technical discussion about potential security
issues. It is not a place for the TOB to have votes about
specification-related business, simply because it is not sane to include
people who are not maintainers of projects in critical security
discussions of said projects.

If in the future we discover that we need to have a place to vote on
security issues, the TOB can do that on their own private mailing list.
For now, we should focus on making sure that security disclosures on
*actual shipping code* is actually done properly.

Signed-off-by: Aleksa Sarai <asarai@suse.de>