generate: replace duplicate env entries
If a given environment variable has already been set, replace the
variable rather than appending it to the environment list. This avoids
issues with POSIX (which doesn't define what to do with duplicate
values).

This required a change in the signature of AddProcessEnv, because doing
the check by splitting around "=" is just not safe (and brings up
questions about variables without "=" in their names that I'd rather
punt on). The check if two variables is the same is if they have the
same NAME= prefix.

Signed-off-by: Aleksa Sarai <asarai@suse.de>