Add T.Skip(count int, description string)
The spec [1] suggests "ok 23 # SKIP ...", but if you attempt this with:

  t.Pass("# SKIP foo")

you get an extra hyphen ("ok 23 - # SKIP foo") which is parsed as a
successful test (and not a skipped test).

Michael gives the following example to motivate 'count' [2]:

  if ConditionIsMet() {
    t.Ok(Foo(), "foo")
    t.Ok(Bar(), "bar")
  } else {
    t.Skip(2, "condition not met")
  }

The spec example of this is in [3], showing no special syntax for
skipping multiple tests (although there is a special syntax for
skipping *all* the tests).

Also alphabetize TESTS in the Makefile.

[1]: http://testanything.org/tap-version-13-specification.html#skipping-tests
[2]: https://github.com/mndrix/tap-go/pull/6#issuecomment-272041306
[3]: http://testanything.org/tap-version-13-specification.html#skipping-a-few