runtime: Semantically version this specification
The command-line interface is largely orthogonal to the config format,
and config authors and runtime callers may be entirely different sets
of people.  Zhang Wei called for more explicit versioning for the CLI
[1].

Wedging multiple, independently versioned entities into a single
repository can be awkward, but earlier proposals to put the CLI in its
own repository [2] were unsuccessful because compliance testing
requires both a CLI and a config specification [3].  I don't think
that's a solid reason [4], but discussion along that line stalled out,
before being revived in today's meeting.  Unfortunately, now this CLI
spec is destined for runtime-tools [5], so we may still have multiple,
independently versioned entities in a single repository.  Wherever
this lands up, it's useful for this CLI spec to be clear about it's
own versioning.

[1]: https://github.com/opencontainers/runtime-spec/pull/513#issuecomment-257204979
[2]: https://github.com/opencontainers/runtime-spec/pull/513#r72818875
[3]: https://github.com/opencontainers/runtime-spec/pull/513#r72866653
[4]: https://github.com/opencontainers/runtime-spec/pull/513#r72868322
[5]: http://ircbot.wl.linuxfoundation.org/meetings/opencontainers/2017/opencontainers.2017-02-08-22.00.log.html#l-125

Signed-off-by: W. Trevor King <wking@tremily.us>