runtime: Replace '$LISTEN_FDS + 3' with '--console-socket FD'
This bridges the gap between my original '--console-socket PATH'
proposal from 15350b31 (runtime: Add --console-socket for terminal
handling, 2017-02-08) and Michael's approach from baa24120 (runtime:
Replace '--console-socket PATH' with '$LISTEN_FDS + 3', 2017-02-08).
All of the benefits of not needing an on-disk socket, and none of the
magic of overloading $LISTEN_FDS.  I don't know why Michael didn't
like this approach as much [1], but it feels a lot better to me.

[1]: http://ircbot.wl.linuxfoundation.org/meetings/opencontainers/2016/opencontainers.2016-10-19-21.00.log.html#l-31

Signed-off-by: W. Trevor King <wking@tremily.us>