runtime: Add compliance language
This language is based on existing language in RFC 2616 [1], RFC 7230
[2], and the runtime spec [3].

By defining both "compliant" and "not compliant" we leave the door
open for a third case where an implementation (which can be a runtime
or a runtime caller) does not violate any MUST-level requirements but
does not allow some MUST-level requirements to be tested.  For
example, a runtime whose 'create' always exited with a non-zero exit
code would not allow for testing the output of 'state' calls, because
the test suite would have no valid container ID to use in the 'state'
calls.  Such a runtime would be neither compliant nor non-compliant.

[1]: https://tools.ietf.org/html/rfc2616#section-1.2
[2]: https://tools.ietf.org/html/rfc7230#section-2.5
[3]: https://github.com/opencontainers/runtime-spec/blob/v1.0.0-rc4/spec.md#notational-conventions

Signed-off-by: W. Trevor King <wking@tremily.us>