validate: Test config.json and rootfs sibling-hood
Test the spec's [1]:

  While these artifacts MUST all be present in a single directory on
  the local filesystem, ...

which is a condition it imposes on config.json and the directory
referenced by root.path.

I think we should drop that restriction from the spec, but my attempt
to remove the restriction was rejected [2].  If a future spec drops
the restriction, we can revert this commit.

Using path/filepath for the path manipulation will break when
validating cross-platform configs (e.g. trying to validate a Windows
bundle on a Linux machine).  But that's a bigger issue than this
commit, so I've left it alone for now.

[1]: https://github.com/opencontainers/runtime-spec/blob/v1.0.0-rc5/bundle.md#container-format
[2]: https://github.com/opencontainers/runtime-spec/pull/469

Signed-off-by: W. Trevor King <wking@tremily.us>