Makefile: Use -set_exit_status in .golint
This landed in golang/lint@58f662d2 (golint: Add a flag to cause a
non-zero exit when suggestions are made, 2016-04-26).  The golint devs
warn about using this in builds to avoid annoying false positives [1],
but we can always add a non-failing form for CI testing if we end up
needing that.

[1]: https://github.com/golang/lint/issues/65#issuecomment-55508544

Signed-off-by: W. Trevor King <wking@tremily.us>