rootfs.tar.gz: Bump to BusyBox 1.25.1
Update from the old 1.24.1 with a new extract from
stage3-amd64-20170907.tar.bz2.  The stage3 had a SHA512 hash of [1]
and a valid signature by [2].

Generated with:

  $ cd contrib/rootfs-builder
  $ touch get-stage3.sh
  $ make rootfs.tar.gz
  $ cp -f rootfs.tar.gz ../..

This change adds the following commands to the rootfs:

* blkdiscard
* nsenter
* ubirename
* unshare

of which nsenter and unshare are interesting for runtime-tools
(although having them *inside* the container probably doesn't matter
much).

[1]: afa5f41a10c9e730cc481d805c30e9f5d1912a30956a23b8780b1e175a5dc764bb8dbf40392e14bfa9a9891b9f380b3287b9b199abce82f83114b8851062efe2
[2]: Gentoo Linux Release Engineering (Automated Weekly Release Key) <releng@gentoo.org>
     Primary key fingerprint: 13EB BDBE DE7A 1277 5DFD  B1BA BB57 2E0E 2D18 2910

Signed-off-by: W. Trevor King <wking@tremily.us>