validation: add a generate smoke-test
To avoid cases where our own validation code would consider our defaults
unsafe (which has happened in the past several times), add a smoke-test
to ensure that this won't happen. Our defaults should not be
intentionally invalid, as that confuses downstreams like umoci which use
runtime-tools for the default as well as for validation of the generated
configuration.

Signed-off-by: Aleksa Sarai <asarai@suse.de>