specerror: Add SplitLevel helper
This doesn't save much time for folks who are iterating over the
removed errors (e.g. to print warnings about them), but Aleksa asked
for it [1] and Ma suggested the helper struct [2].

I haven't used it in runtimetest.  I'm waiting for the in-flight
7aa3987 (cmd/runtimetest/main: Use TAP diagnostics for errors,
2017-07-28, #439) to settle, since that commit cleans up some of the
runtimetest error-processing code with:

  validations := defaultValidations
  if platform == "linux" {
    validations = append(validations, linuxValidations...)
  }

and a single loop over the constructed validations array.

[1]: https://github.com/opencontainers/runtime-tools/pull/492#discussion_r142488597
[2]: https://github.com/opencontainers/runtime-tools/pull/492#discussion_r143620920

Signed-off-by: W. Trevor King <wking@tremily.us>