runtimetest: Raise ConfigInRootBundleDir for missing config.json
This is a bit tricky, because before we load the config, we don't know
the target spec version.  If a future spec version relaxes the current
requirement, we'll have difficulty reporting an "I couldn't find the
config" error as a spec violation.  If/when that happens, I'm fine
dropping this block altogether, and having this error fall through as
a non-spec "you called us wrong" error.

Signed-off-by: W. Trevor King <wking@tremily.us>