runtimetest: Make validateRlimits silent on Windows
This isn't a question of not being implemented yet.  Windows does not,
and is unlikely to ever, support rlimits [1].  And the spec only
defines process.rlimits for POSIX platforms [2].  Folks writing
Windows configs will only be setting process.rlimits as an out-of-spec
extention.  We, as runtime validators, are unlikely to write such
configs.  But if we use runtimetest in a container launched from such
a config, we should silently ignore that out-of-spec extention, just
like we silently ignore all other out-of-spec extentions, instead of
logging a warning.

[1]: https://github.com/opencontainers/runtime-spec/pull/835#issuecomment-303455386
[2]: https://github.com/opencontainers/runtime-spec/blob/v1.0.0/config.md#posix-process

Signed-off-by: W. Trevor King <wking@tremily.us>