filepath/abs_test: Compare IsAbs with the standard library
For the test cases that match the current GOOS.  This ensures we stay
consistent (at least for platforms where we run tests).

Also drop an unecessary trailing newline from the old error message.

I don't compare for Abs, because the stdlib doesn't have a
configurable cwd.  We could check the current working directory
[1,2,...], but I don't want to use the frozen syscall package and am
not interested enough to import lots of per-OS x/sys packages.  Also,
the odds of the test working directory matching one of the cwd values
(or even of those cwd values actually existing on the host) are small,
and it would be even more work to temporarily change to (and possibly
create) those directories per test.

[1]: https://golang.org/pkg/syscall/#Getcwd
[2]: https://godoc.org/golang.org/x/sys/unix#Getcwd

Signed-off-by: W. Trevor King <wking@tremily.us>